+++ Sources/ENCore/Common/ExposureManager/Models/ExposureConfiguration.swift
@@ -9,9 +9,9 @@ import Foundation
 
 protocol ExposureConfiguration {
     var minimumRiskScope: UInt8 { get }
-    var attenuationLevelValues: [Int] { get }
-    var daysSinceLastExposureLevelValues: [Int] { get }
-    var durationLevelValues: [Int] { get }
-    var transmissionRiskLevelValues: [Int] { get }
+    var attenuationLevelValues: [UInt8] { get }
+    var daysSinceLastExposureLevelValues: [UInt8] { get }
+    var durationLevelValues: [UInt8] { get }
+    var transmissionRiskLevelValues: [UInt8] { get }
     var attenuationDurationThresholds: [Int] { get }
 }
